update to 4o-audio-preview, bugfixes, and commentary
Moved from whisper to 4o-audio-preview, fixed a moviepy import error, and overall small changes in the commentary to go along with model changes.  Additionally re-ran the notebook to ensure cell data was correct.

update to 4o-audio-preview, bugfixes, and commentary

Moved from whisper to 4o-audio-preview, fixed a moviepy import error, and overall small changes in the commentary to go along with model changes